questionservice/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
pymongo==4.6.0
lxml==5.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
